climbnet/requirements.txt
blinker==1.8.2
click==8.1.7
filelock==3.15.1
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.0
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
pandas==2.2.2
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.3.1
tzdata==2024.1
Werkzeug==3.0.3
